refactor(ui): thin settings and workflow route files
